Combined kelp-toc and kelp-toc-nested
