+++ js/kelp.js
@@ -1,5 +1,4 @@
 import './components/toc.js';
-import './components/toc-nested.js';
 import './components/heading-anchors.js';
 import './components/toggle-pw.js';
 import './components/autogrow.js';
