+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "kelpui",
-	"version": "0.13.1",
+	"version": "0.14.0",
 	"description": "A UI library for people who love HTML, powered by modern CSS and Web Components.",
 	"keywords": [
 		"html",
