--- tests/toc-nested/default.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<title>Kelp</title>
-		<meta name="description" content="A UI library for people who love HTML, powered by modern CSS and Web Components.">
-
-		<!-- Mobile Screen Resizing -->
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-
-		<!-- Kelp CSS -->
-		<link rel="stylesheet" type="text/css" href="../../css/kelp.css">
-
-		<!-- Kelp JS -->
-		<script type="module" src="../../js/kelp.js"></script>
-		<script>
-			// Ensures we can test the ready event
-			document.addEventListener('kelp:tocnested-ready', (event) => {
-				console.log('ready');
-			});
-		</script>
-	</head>
-	<body>
-
-		<main class="container">
-
-			<h1>kelp-toc-nested - default settings</h1>
-
-			<kelp-toc-nested hide-until-ready></kelp-toc-nested>
-
-			<div id="content">
-				<h2>Cat O'Nine Tails</h2>
-				<h3 id="the-brig">The Brig</h3>
-				<h4>Privateer</h4>
-				<h2>Ahoy</h2>
-				<h3 id="man-of-war">Man-of-War</h3>
-				<h2 id="corsair">Corsair</h2>
-				<h3 id="shiver-me-timbers">Shiver Me Timbers</h3>
-				<h2 id="scurvy-dog">Scurvy Dog</h2>
-				<h4>Sea Legs</h4>
-				<h2 id="quarterdeck">On the Quarterdeck</h2>
-				<h3 id="jolly-roger">Jolly Roger</h3>
-				<h3 id="sloop">Sloop</h3>
-				<h4 id="swab">Swab</h4>
-				<h5>Grog</h5>
-			</div>
-
-		</main>
-
-	</body>
-</html>
